Enum air_test_framework::asserts::ServiceTagName

pub enum ServiceTagName {
    Ok,
    Error,
    SeqOk,
    SeqError,
    Behaviour,
    DbgBehaviour,
    Map,
}

Auto-generated discriminant enum variants

Variants

Ok

Simple service that returns same value

Error

Simple service that returns same call result (i.e. may return a error)

SeqOk

Service that may return a new value on subsequent call. Its keys are either call number string starting from โ0โ, or โdefaultโ.

SeqError

Behaviour

Some known service by name: โechoโ, โunitโ (more to follow).

DbgBehaviour

Same services as defined by the enum element above with dbg! applied to the arguments.

Map

Maps first argument to a value