Struct air_test_framework::execution::AirScriptExecutor

pub struct AirScriptExecutor<R = DefaultAirRunner> { /* private fields */ }

A executor for an AIR script. Several executors may share same TransformedAirScript and its state.

Implementations

impl AirScriptExecutor<DefaultAirRunner>

pub async fn from_annotated( test_parameters: TestRunParameters, annotated_air_script: &str, ) -> Result<Self, String>

Simple constructor where everything is generated from the annotated_air_script.

impl<R: AirRunner> AirScriptExecutor<R>

pub async fn from_transformed_air_script( test_parameters: TestRunParameters, test_init_parameters: TestInitParameters, transformed_air_script: TransformedAirScript<R>, ) -> Result<Self, String>

pub async fn new( test_parameters: TestRunParameters, common_services: Vec<MarineServiceHandle>, extra_peers: impl IntoIterator<Item = PeerId>, annotated_air_script: &str, ) -> Result<Self, String>

Create execution from the annotated air script.

extra_peers allows you to define peers that are not mentioned in the annotated script explicitly, but are used, e.g. if their names are returned from a call.

pub async fn from_network( test_parameters: TestRunParameters, test_init_parameters: TestInitParameters, network: Rc<Network<R>>, annotated_air_script: &str, ) -> Result<Self, String>

pub async fn new_with_init_parameters( test_parameters: TestRunParameters, test_init_parameters: TestInitParameters, common_services: Vec<MarineServiceHandle>, extra_peers: impl IntoIterator<Item = PeerId>, annotated_air_script: &str, ) -> Result<Self, String>

pub fn execution_iter<'s, Id>( &'s self, peer_id: &Id, ) -> Option<impl Stream<Item = RawAVMOutcome> + 's>
where PeerId: Borrow<Id> + for<'a> From<&'a Id>, Id: Eq + Hash + ?Sized,

Return Iterator for handling all the queued datas for particular peer_id.

pub async fn execute_all<Id>(&self, peer_id: &Id) -> Option<Vec<RawAVMOutcome>>
where PeerId: Borrow<Id> + for<'a> From<&'a Id>, Id: Eq + Hash + ?Sized,

Process all queued datas, panicing on error.

pub async fn execute_one<Id>(&self, peer_id: &Id) -> Option<RawAVMOutcome>
where PeerId: Borrow<Id> + for<'a> From<&'a Id>, Id: Eq + Hash + ?Sized,

Process one queued data, panicing if it is unavalable or on error.

pub fn enqueue(&self, peer_id: impl Into<PeerId>, data: Data)

Push data into peer’s queue.

pub fn get_prev_data(&self, peer_id: impl Into<PeerId>) -> Data

pub fn resolve_name(&self, name: &str) -> PeerId

pub fn get_network(&self) -> Rc<Network<R>>

pub fn get_transformed_air_script(&self) -> &TransformedAirScript<R>